ml/requirements.txt
scikit-learn>=1.3.0
imbalanced-learn>=0.11.0
xgboost>=2.0.0
shap>=0.43.0
matplotlib>=3.7.0
seaborn>=0.13.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0
pyyaml>=6.0
